Copy Measure Reference from Execute Task to MeasureReport
Before the url property of the measure resource was used but did contain
an invalid reference created with the transaction bundle internal
reference id and never got updated to the actual measure resource id in
the FHIR store. This change lets the measure report contain a valid
reference to the measure resource.